refactor(cli): moved to own commands
